+++ app/main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from app.routers import reservation, service_site, queue_arranging, authentication
+from app.routers import reservation, service_site, queue_arranging, authentication, report
 from fastapi.middleware.cors import CORSMiddleware
 
 description = """
@@ -24,7 +24,7 @@ tags_metadata = [
 app = FastAPI(
     title="Service Site OGYH",
     description=description,
-    version="0.1",
+    version="1.0",
     openapi_tags=tags_metadata,
 )
 
@@ -40,7 +40,13 @@ app.include_router(reservation.router, prefix="/api")
 app.include_router(service_site.router, prefix="/api")
 app.include_router(queue_arranging.router, prefix="/api")
 app.include_router(authentication.router)
+app.include_router(report.router)
 
+@app.on_event("startup")
+async def startup_event():
+    import requests
+    res = requests.post("https://wcg-apis-test.herokuapp.com/register_user", params={"username":"Chayapol", "password": "Kp6192649"})
+    print(res.status_code)
 
 @app.get("/", include_in_schema=False)
 async def read_root():
