+++ app/routers/reservation.py
@@ -6,6 +6,7 @@ from app.utils.utils import (
     arranging_reservation_by_site_name,
     fetch_url,
     get_service_site_avaliable,
+    get_access_to_api
 )
 from app.utils.paginator import Paginator
 from app.database import retrieve_site
@@ -13,8 +14,6 @@ from app.models.basic_model import Message
 from app.models.reservation import (
     GetReservationsResponse,
     GetReservationResponse,
-    example_get_reservations,
-    example_reservation,
 )
 import bson
 import requests
@@ -27,12 +26,6 @@ router = APIRouter(prefix="/site/{site_id}", tags=["vaccine reservation"])
     response_description="Reservation retrived",
     summary="Get every reservations",
     response_model=GetReservationsResponse,
-    responses={
-        status.HTTP_200_OK: {
-            "description": "Found a service site",
-            "content": {"application/json": {"example": example_get_reservations}},
-        }
-    },
 )
 async def read_users_reservations(
     site_id: str, limit: Optional[int] = None, page: Optional[int] = 1
@@ -43,12 +36,8 @@ async def read_users_reservations(
     - **limit** : number of users to be shown as a result
     - **page** : number of pages to be shown as a result
     """
-    res = requests.post(
-        "https://wcg-apis-test.herokuapp.com/login", auth=("Chayapol", "Kp6192649")
-    )
-    access_token = res.json()["access_token"]
     user_data = fetch_url(
-        "https://wcg-apis-test.herokuapp.com/reservations", token=access_token
+        "https://wcg-apis-test.herokuapp.com/reservations", token=get_access_to_api()
     )
     user_data_by_site_name = arranging_reservation_by_site_name(user_data)
     try:
@@ -81,10 +70,6 @@ async def read_users_reservations(
     response_model=GetReservationResponse,
     responses={
         status.HTTP_404_NOT_FOUND: {"model": Message, "description": "Not found"},
-        status.HTTP_200_OK: {
-            "description": "Found a reservations",
-            "content": {"application/json": {"example": example_reservation}},
-        },
     },
 )
 async def read_users_reservation(
@@ -98,21 +83,15 @@ async def read_users_reservation(
     - **citizen_id**: a specific citizen_id
 
     """
-    res = requests.post(
-        "https://wcg-apis-test.herokuapp.com/login", auth=("Chayapol", "Kp6192649")
-    )
-    access_token = res.json()["access_token"]
     user_data = fetch_url(
-        "https://wcg-apis-test.herokuapp.com/reservations", token=access_token
+        "https://wcg-apis-test.herokuapp.com/reservations", token=get_access_to_api()
     )
     user_data_by_site_name = arranging_reservation_by_site_name(user_data)
-
     try:
         site = await retrieve_site(site_id)
     except bson.errors.InvalidId:
         message = f"Service site id {site_id} is invalid"
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=message)
-    print(site)
     if site:
         search_site = get_service_site_avaliable(
             data=user_data_by_site_name, key=site["name"]
